--- _projects/games/mansionGame/js/MansionLogic/Background.js
@@ -1,123 +1,0 @@
-import GameObject from './GameObject.js';
-
-/** Background GameObject
- */
-export class Background extends GameObject {
-    /**
-     * Constructor is called by GameLevel create() method
-     * @param {Object} data - The data object for the background
-     * @param {Object} gameEnv - The game environment object for convenient access to game properties 
-     */
-    constructor(data = null, gameEnv = null) {
-        super(gameEnv);
-
-        if (!data.src) {
-            throw new Error('Background requires a src property in data');
-        }
-        this.data = data;
-        // Set the properties of the background
-        this.image = new Image();
-        this.image.src = data.src;
-        this.isInitialized = false; // Flag to track initialization
-
-        // Finish initializing the background after the image loads 
-        this.image.onload = () => {
-            // Width and height come from the image
-            this.width = this.image.width;
-            this.height = this.image.height;
-
-            // Create the canvas element and context
-            this.canvas = document.createElement("canvas");
-            this.canvas.style.position = "absolute";
-            this.canvas.style.zIndex = this.data.zIndex || "0";
-            this.canvas.id = data.id || "background";
-            this.ctx = this.canvas.getContext("2d", { willReadFrequently: true });
-            
-            // Align the canvas size to the gameCanvas
-            this.alignCanvas();
-
-            // Append the canvas to the DOM
-            const parent = this.gameEnv?.gameContainer || document.getElementById("gameContainer") || this.gameEnv?.canvas?.parentElement;
-            if (!parent) {
-                throw new Error(`Cannot create ${this.canvas.id}: missing Mansion Game container`);
-            }
-            parent.appendChild(this.canvas);
-            this.isInitialized = true; // Mark as initialized
-        };
-    }
-
-    /**
-     * Align canvas to be the same size and position as the gameCanvas 
-     */
-    alignCanvas() {
-        // align the canvas to the gameCanvas, Layered
-        const gameCanvas = this.gameEnv?.canvas || this.gameEnv?.gameCanvas || document.getElementById("gameCanvas");
-        if (!gameCanvas) {
-            console.error("Game canvas not found");
-            return;
-        }
-        this.canvas.width = gameCanvas.width;
-        this.canvas.height = gameCanvas.height;
-        this.canvas.style.width = `${gameCanvas.width}px`;
-        this.canvas.style.height = `${gameCanvas.height}px`;
-        this.canvas.style.left = gameCanvas.style.left;
-        this.canvas.style.top = gameCanvas.style.top;
-    }
-
-    /**
-     * Update is called by GameLoop on all GameObjects 
-     */
-    update() {
-        // Draw the background image
-        this.draw();
-    }
-
-    /**
-     * Draws the background image within the canvas 
-     */
-    draw() {
-        if (!this.isInitialized) {
-            return; // Skip drawing if not initialized
-        }
-
-        // Clear the canvas
-        const canvasWidth = this.canvas.width;
-        const canvasHeight = this.canvas.height;
-        this.ctx.clearRect(0, 0, canvasWidth, canvasHeight);
-        
-        if (this.image) {
-            // Draw the background image
-            this.ctx.drawImage(this.image, 0, 0, canvasWidth, canvasHeight);
-        } else {
-            // Default fill color if no image is provided
-            this.ctx.fillStyle = this.data.color || '#242435';
-            this.ctx.fillRect(0, 0, canvasWidth, canvasHeight);
-        }
-    }
-    
-    /**
-     * Resize method is called by resize listner on all GameObjects
-     */
-    resize() {
-        this.alignCanvas(); // Align the canvas to the gameCanvas
-        this.draw(); // Redraw the canvas after resizing
-    }
-    
-    /**
-     * Destroy method is called to clean up resources
-     */
-    destroy() {
-        if (this.canvas && this.canvas.parentNode) {
-            this.canvas.parentNode.removeChild(this.canvas);
-        }
-        
-        if (this.gameEnv && this.gameEnv.gameObjects) {
-            const index = this.gameEnv.gameObjects.indexOf(this);
-            if (index !== -1) {
-                this.gameEnv.gameObjects.splice(index, 1);
-            }
-        }
-    }
-}
-
-export default Background;
