--- _projects/games/mansionGame/js/MansionLogic/Player.js
@@ -1,209 +1,0 @@
-import Character from './Character.js';
-import TouchControls from './TouchControls.js';
-
-// Define non-mutable constants as defaults
-const SCALE_FACTOR = 25; // 1/nth of the height of the canvas
-const STEP_FACTOR = 100; // 1/nth, or N steps up and across the canvas
-const ANIMATION_RATE = 1; // 1/nth of the frame rate
-const INIT_POSITION = { x: 0, y: 0 };
-
-
-class Player extends Character {
-    // Static counter for unique player IDs (uninitialized)
-    static playerCount;
-    /**
-     * The constructor method is called when a new Player object is created.
-     * 
-     * @param {Object|null} data - The sprite data for the object. If null, a default red square is used.
-     */
-    constructor(data = null, gameEnv = null) {
-        super(data, gameEnv);
-        // Increment static player counter and assign unique id
-        Player.playerCount = (Player.playerCount || 0) + 1;
-        this.id = data?.id ? data.id.toLowerCase() : `player${Player.playerCount}`;
-        this.keypress = data?.keypress || {up: 87, left: 65, down: 83, right: 68};
-        this.touchOptions = data?.touchOptions || {interactLabel: "e", position: "left"};
-        this.touchOptions.id = `touch-controls-${this.id}`;
-        this.touchOptions.mapping = this.keypress;
-        this.pressedKeys = {}; // active keys array
-        this.bindMovementKeyListners();
-        this.gravity = data.GRAVITY || false;
-        this.acceleration = 0.001;
-        this.time = 0;
-        this.moved = false;
-        // Initialize touch controls for mobile devices
-        this.touchControls = new TouchControls(gameEnv, this.touchOptions);
-    }
-
-    /**
-     * Binds key event listeners to handle object movement.
-     * 
-     * This method binds keydown and keyup event listeners to handle object movement.
-     * The .bind(this) method ensures that 'this' refers to the object object.
-     */
-    bindMovementKeyListners() {
-        addEventListener('keydown', this.handleKeyDown.bind(this));
-        addEventListener('keyup', this.handleKeyUp.bind(this));
-    }
-
-    handleKeyDown({ keyCode }) {
-        // capture the pressed key in the active keys array
-        this.pressedKeys[keyCode] = true;
-        // set the velocity and direction based on the newly pressed key
-        this.updateVelocityAndDirection();
-    }
-
-    /**
-     * Handles key up events to stop the player's velocity.
-     * 
-     * This method stops the player's velocity based on the key released.
-     * 
-     * @param {Object} event - The keyup event object.
-     */
-    handleKeyUp({ keyCode }) {
-        // remove the lifted key from the active keys array
-        if (keyCode in this.pressedKeys) {
-            delete this.pressedKeys[keyCode];
-        }
-        // adjust the velocity and direction based on the remaining keys
-        this.updateVelocityAndDirection();
-    }
-
-    /**
-     * Update the player's velocity and direction based on the pressed keys.
-     */
-    updateVelocityAndDirection() {
-        this.velocity.x = 0;
-        this.velocity.y = 0;
-        const xVel = this.xVelocity * 0.7;
-
-        // Multi-key movements (diagonals: upLeft, upRight, downLeft, downRight)
-        if (this.pressedKeys[this.keypress.up] && this.pressedKeys[this.keypress.left]) {
-            this.velocity.y -= this.yVelocity;
-            this.velocity.x -= xVel;
-            this.direction = 'upLeft';
-        } else if (this.pressedKeys[this.keypress.up] && this.pressedKeys[this.keypress.right]) {
-            this.velocity.y -= this.yVelocity;
-            this.velocity.x += xVel;
-            this.direction = 'upRight';
-        } else if (this.pressedKeys[this.keypress.down] && this.pressedKeys[this.keypress.left]) {
-            this.velocity.y += this.yVelocity;
-            this.velocity.x -= xVel;
-            this.direction = 'downLeft';
-        } else if (this.pressedKeys[this.keypress.down] && this.pressedKeys[this.keypress.right]) {
-            this.velocity.y += this.yVelocity;
-            this.velocity.x += xVel;
-            this.direction = 'downRight';
-        // Single key movements (left, right, up, down) 
-        } else if (this.pressedKeys[this.keypress.up]) {
-            this.velocity.y -= this.yVelocity;
-            this.direction = 'up';
-            this.moved = true;
-        } else if (this.pressedKeys[this.keypress.left]) {
-            this.velocity.x -= xVel;
-            this.direction = 'left';
-            this.moved = true;
-        } else if (this.pressedKeys[this.keypress.down]) {
-            this.velocity.y += this.yVelocity;
-            this.direction = 'down';
-            this.moved = true;
-        } else if (this.pressedKeys[this.keypress.right]) {
-            this.velocity.x += xVel;
-            this.direction = 'right';
-            this.moved = true;
-        } else{
-            this.moved = false;
-        }
-    }
-    update() {
-        super.update();
-        if(!this.moved){
-            if (this.gravity) {
-                    this.time += 1;
-                    this.velocity.y += 0.5 + this.acceleration * this.time;
-                }
-            }
-        else{
-            this.time = 0;
-        }
-        }
-        
-    /**
-     * Overrides the reaction to the collision to handle
-     *  - clearing the pressed keys array
-     *  - stopping the player's velocity
-     *  - updating the player's direction   
-     * @param {*} other - The object that the player is colliding with
-     */
-    handleCollisionReaction(other) {    
-        this.pressedKeys = {};
-        this.updateVelocityAndDirection();
-        super.handleCollisionReaction(other);
-    }
-
-    /**
-     * Toggle touch controls visibility (useful for debugging or user preference)
-     */
-    toggleTouchControls() {
-        if (this.touchControls) {
-            this.touchControls.toggle();
-        }
-    }
-
-    /**
-     * Show touch controls explicitly
-     */
-    showTouchControls() {
-        if (this.touchControls) {
-            this.touchControls.show();
-        }
-    }
-
-    /**
-     * Hide touch controls explicitly  
-     */
-    hideTouchControls() {
-        if (this.touchControls) {
-            this.touchControls.hide();
-        }
-    }
-
-    /**
-     * Show the interact button when near an NPC
-     */
-    showInteractButton() {
-        if (this.touchControls) {
-            this.touchControls.showInteractButton();
-        }
-    }
-
-    /**
-     * Hide the interact button when not near an NPC
-     */
-    hideInteractButton() {
-        if (this.touchControls) {
-            this.touchControls.hideInteractButton();
-        }
-    }
-
-    /**
-     * Check if interact button is currently visible
-     */
-    isInteractButtonVisible() {
-        return this.touchControls ? this.touchControls.isInteractButtonVisible() : false;
-    }
-
-    /**
-     * Clean up resources when player is destroyed
-     */
-    destroy() {
-        if (this.touchControls) {
-            this.touchControls.destroy();
-        }
-        super.destroy?.();
-    }
-
-
-}
-
-export default Player;
