+++ _projects/games/mansionGame/levels/mansionLevel4.js
@@ -1,13 +1,13 @@
 // To build GameLevels, each contains GameObjects from below imports
-import GameEnvBackground from './MansionLogic/GameEnvBackground.js';
-import Player from './MansionLogic/Player.js';
-import GameObject from './MansionLogic/GameObject.js';
-import DialogueSystem from './MansionLogic/DialogueSystem.js';
+import GameEnvBackground from '@assets/js/GameEnginev1.1/essentials/GameEnvBackground.js';
+import Player from '@assets/js/GameEnginev1.1/essentials/Player.js';
+import GameObject from '@assets/js/GameEnginev1.1/essentials/GameObject.js';
+import DialogueSystem from '@assets/js/GameEnginev1.1/essentials/DialogueSystem.js';
 
 //Import custom classes from select files
-import Barrier from './CustomGameClasses/Barrier.js';
-import BlackjackGameManager from './CustomGameClasses/Blackjack.js';
-import TriggerZone from './CustomGameClasses/TriggerZone.js';
+import Barrier from './Barrier.js';
+import BlackjackGameManager from './Blackjack.js';
+import TriggerZone from './TriggerZone.js';
 
 class MansionLevel4 {
     constructor(gameEnv) {
