+++ _projects/games/mansionGame/levels/scoreSettings.js
